# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * DevPulse API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TaskColumn = typeof TaskColumn[keyof typeof TaskColumn];


export const TaskColumn = {
  todo: 'todo',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  column: TaskColumn;
  priority: TaskPriority;
  /** @nullable */
  tags?: string | null;
  position?: number;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export type TaskInputColumn = typeof TaskInputColumn[keyof typeof TaskInputColumn];


export const TaskInputColumn = {
  todo: 'todo',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export type TaskInputPriority = typeof TaskInputPriority[keyof typeof TaskInputPriority];


export const TaskInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  column: TaskInputColumn;
  priority: TaskInputPriority;
  tags?: string;
  position?: number;
}

export type TaskUpdateColumn = typeof TaskUpdateColumn[keyof typeof TaskUpdateColumn];


export const TaskUpdateColumn = {
  todo: 'todo',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export type TaskUpdatePriority = typeof TaskUpdatePriority[keyof typeof TaskUpdatePriority];


export const TaskUpdatePriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  column?: TaskUpdateColumn;
  priority?: TaskUpdatePriority;
  tags?: string;
  position?: number;
}

export interface TaskStats {
  todo: number;
  inProgress: number;
  completed: number;
  total: number;
}

export interface FavoriteDeveloper {
  id: number;
  githubLogin: string;
  avatarUrl: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  followers?: number | null;
  /** @nullable */
  publicRepos?: number | null;
  savedAt: string;
}

export interface FavoriteDeveloperInput {
  githubLogin: string;
  avatarUrl: string;
  name?: string;
  bio?: string;
  followers?: number;
  publicRepos?: number;
}