# lib/api-zod/src/generated/types/favoriteDeveloper.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * DevPulse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FavoriteDeveloper {
  id: number;
  githubLogin: string;
  avatarUrl: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  followers?: number | null;
  /** @nullable */
  publicRepos?: number | null;
  savedAt: string;
}