# lib/api-zod/src/generated/types/favoriteDeveloperInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * DevPulse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FavoriteDeveloperInput {
  githubLogin: string;
  avatarUrl: string;
  name?: string;
  bio?: string;
  followers?: number;
  publicRepos?: number;
}