# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * DevPulse API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskColumn } from './taskColumn';
import type { TaskPriority } from './taskPriority';

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  column: TaskColumn;
  priority: TaskPriority;
  /** @nullable */
  tags?: string | null;
  position?: number;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}