# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * DevPulse API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskInputColumn } from './taskInputColumn';
import type { TaskInputPriority } from './taskInputPriority';

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  column: TaskInputColumn;
  priority: TaskInputPriority;
  tags?: string;
  position?: number;
}