# lib/api-zod/src/generated/types/taskStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * DevPulse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TaskStats {
  todo: number;
  inProgress: number;
  completed: number;
  total: number;
}