# lib/api-zod/src/generated/types/taskUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * DevPulse API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskUpdateColumn } from './taskUpdateColumn';
import type { TaskUpdatePriority } from './taskUpdatePriority';

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  column?: TaskUpdateColumn;
  priority?: TaskUpdatePriority;
  tags?: string;
  position?: number;
}